push

+++ app/Http/Controllers/Admin/UsersController.php
@@ -13,59 +13,59 @@ use Illuminate\Support\Facades\Auth;
 class UsersController extends Controller
 {
     public function index(Request $request)
-{
+    {
 
-    $users = User::when($request->search, function ($query, $search) {
+        $users = User::when($request->search, function ($query, $search) {
             return $query->where('name', 'like', '%' . $search . '%');
         })
-        ->when($request->created_at, function ($query, $date) {
-            return $query->whereDate('created_at', $date);
-        })
-        ->paginate(1);
+            ->when($request->created_at, function ($query, $date) {
+                return $query->whereDate('created_at', $date);
+            })
+            ->paginate(1);
 
-    // Trả về view với dữ liệu đã phân trang
-    return view('admin.users.index', compact('users'));
-}
+        // Trả về view với dữ liệu đã phân trang
+        return view('admin.users.index', compact('users'));
+    }
 
 
 
     public function create()
     {
-        $roles = Role::whereIn('id', [1,2,3])->get();
+        $roles = Role::whereIn('id', [1, 2, 3])->get();
         return view('admin.users.create', compact('roles'));
     }
 
     public function store(Request $request)
-{
-    // Validate form inputs
-    $validatedData = $request->validate([
-        'name' => 'required|string|max:255',
-        'email' => 'required|string|email|max:255|unique:users',
-        'role_id' => 'required|integer'
-    ]);
-
-    // dd($validatedData);
-
-    // Tạo người dùng mới với mật khẩu được mã hoá
-    try{
-        User::create([
-            'name' => $request->input('name'),
-            'email' => $request->input('email'),
-            'role_id' => $request->input('role_id'),
-            'password' => Hash::make('12345678'), // Hash password
+    {
+        // Validate form inputs
+        $validatedData = $request->validate([
+            'name' => 'required|string|max:255',
+            'email' => 'required|string|email|max:255|unique:users,email',
+            'role_id' => 'required|integer'
+        ], [
+            'email.unique' => __('messages.email_already_exists') // Custom error message
         ]);
 
-        return redirect()->route('admin.users.index')->with('success', 'User created successfully!');
+        // dd($validatedData);
 
-    }catch(\Exception $e){
-        return redirect()->back()->with('error', 'Failed to add user!');
+        // Tạo người dùng mới với mật khẩu được mã hoá
+        try {
+            User::create([
+                'name' => $request->input('name'),
+                'email' => $request->input('email'),
+                'role_id' => $request->input('role_id'),
+                'password' => Hash::make('12345678'), // Hash password
+            ]);
+
+            return redirect()->route('admin.users.index')->with('success', __('messages.success_created_user'));
+        } catch (\Exception $e) {
+            return redirect()->back()->with('error', 'Failed to add user!');
+        }
     }
-
-}
 
     public function edit(User $user)
     {
-        $roles = Role::whereIn('id', [1,2,3])->get();
+        $roles = Role::whereIn('id', [1, 2, 3])->get();
         return view('admin.users.edit', compact('user', 'roles'));
     }
 
@@ -78,7 +78,7 @@ class UsersController extends Controller
             'role_id' => 'required|integer|in:1,2,3',
         ]);
 
-        try{
+        try {
             $user->name = $request->input('name');
             $user->email = $request->input('email');
             $user->role_id = $request->input('role_id');
@@ -89,24 +89,22 @@ class UsersController extends Controller
             }
 
             $user->save();
-            return redirect()->route('admin.users.index')->with('success', 'User updated successfully.');
-        }
-        catch(\Exception $e){
+            return redirect()->route('admin.users.index')->with('success', __('messages.success_updated_user'));
+        } catch (\Exception $e) {
             return redirect()->back()->with('error', 'Failed to updated user!');
         }
-
-
     }
 
     public function destroy(User $user)
     {
-        try{
+        try {
+            if ($user->tasks()->exists()) {
+                return redirect()->back()->with('error', __('messages.cannot_delete_user_with_tasks'));
+            }
             $user->delete();
-            return redirect()->route('admin.users.index')->with('success', 'User deleted successfully.');
-        }
-        catch(\Exception $e){
+            return redirect()->route('admin.users.index')->with('success', __('messages.success_deleted_user'));
+        } catch (\Exception $e) {
             return redirect()->back()->with('error', 'Failed to delete user!');
         }
-
     }
 }
